Cleaning up the code and eliminating duplicates
Hi Francesco, nice work on this! It is just a little bit difficult to read because right now your code lacks structure, and there is a lot of duplicate code. How about using methods for code that seems to be used a lot? This will improve the structure. And how about adding a switch case statement? This will make it easy to read so that there aren't too many ifs to go through. We will discuss this in class.